mactry/bangladoc_surya_clean/backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "bangladoc"
version = "1.0.0"
description = "Bangla + English PDF OCR with Surya-first pipeline"
requires-python = ">=3.11"
dependencies = [
    "surya-ocr>=0.6.0",
    "transformers<5",
    "PyMuPDF>=1.24.0",
    "pdfplumber>=0.11.0",
    "Pillow>=10.0.0",
    "opencv-python-headless>=4.8.0",
    "easyocr>=1.7.0",
    "google-genai>=1.0.0",
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "python-multipart>=0.0.9",
    "aiofiles>=23.0",
    "python-dotenv>=1.0",
    "numpy>=1.26",
    "pandas>=2.0",
    "pyarrow>=16.0",
    "requests>=2.31",
    "slowapi>=0.1.9",
    "httpx>=0.27",
]

[project.optional-dependencies]
dev = ["pytest>=8.0", "pytest-asyncio>=0.23", "httpx>=0.27", "ruff>=0.4"]

[project.scripts]
bangladoc = "bangladoc_ocr.cli:main"

[tool.ruff]
line-length = 100
target-version = "py311"
